Refactored test_endpoints.py into a folder containing test files per endpoint
