+++ api/app/signals/apps/questionnaires/tests/views/test_public_questions_endpoint.py
@@ -2,7 +2,6 @@
 # Copyright (C) 2021 Gemeente Amsterdam
 import os
 
-from django.contrib.auth import get_user_model
 from django.test import override_settings
 from django.urls import include, path
 from django.utils import timezone
@@ -35,135 +34,6 @@ test_urlconf.urlpatterns = urlpatterns
 
 
 @override_settings(ROOT_URLCONF=test_urlconf)
-class TestPublicQuestionnaireEndpoint(ValidateJsonSchemaMixin, APITestCase):
-    base_endpoint = '/public/qa/questionnaires/'
-
-    def setUp(self):
-        self.questionnaire = QuestionnaireFactory.create()
-
-        self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_get_questionnaire_detail.json')
-        )
-        self.list_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_get_questionnaire_list.json')
-        )
-
-    def test_questionnaire_list(self):
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.list_schema, response.json())
-
-    def test_questionnaire_detail_by_key(self):
-        response = self.client.get(f'{self.base_endpoint}{self.questionnaire.uuid}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.detail_schema, response.json())
-
-    def test_questionnaire_detail_by_uuid(self):
-        response = self.client.get(f'{self.base_endpoint}{self.questionnaire.uuid}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.detail_schema, response.json())
-
-    def test_questionnaire_create_not_allowed(self):
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 405)
-
-    def test_questionnaire_update_not_allowed(self):
-        response = self.client.patch(f'{self.base_endpoint}{self.questionnaire.uuid}', data={})
-        self.assertEqual(response.status_code, 405)
-
-    def test_questionnaire_delete_not_allowed(self):
-        response = self.client.delete(f'{self.base_endpoint}{self.questionnaire.uuid}')
-        self.assertEqual(response.status_code, 405)
-
-
-@override_settings(ROOT_URLCONF=test_urlconf)
-class TestPrivateQuestionnaireEndpoint(ValidateJsonSchemaMixin, APITestCase):
-    base_endpoint = '/private/qa/questionnaires/'
-
-    def setUp(self):
-        user_model = get_user_model()
-        self.superuser, _ = user_model.objects.get_or_create(
-            email='signals.admin@example.com', is_superuser=True,
-            defaults={'first_name': 'John', 'last_name': 'Doe', 'is_staff': True}
-        )
-        self.questionnaire = QuestionnaireFactory.create()
-
-        self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/private_get_questionnaire_detail.json')
-        )
-        self.list_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/private_get_questionnaire_list.json')
-        )
-
-    def test_questionnaire_list(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.list_schema, response.json())
-
-        self.client.logout()
-
-    def test_questionnaire_detail(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.get(f'{self.base_endpoint}{self.questionnaire.id}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.detail_schema, response.json())
-
-        self.client.logout()
-
-    def test_questionnaire_list_unauthorized(self):
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_questionnaire_detail_unauthorized(self):
-        response = self.client.get(f'{self.base_endpoint}{self.questionnaire.id}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_questionnaire_create_unauthorized(self):
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 401)
-
-    def test_questionnaire_update_unauthorized(self):
-        response = self.client.patch(f'{self.base_endpoint}{self.questionnaire.id}', data={})
-        self.assertEqual(response.status_code, 401)
-
-    def test_questionnaire_delete_unauthorized(self):
-        response = self.client.delete(f'{self.base_endpoint}{self.questionnaire.id}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_questionnaire_create_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-    def test_questionnaire_update_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.patch(f'{self.base_endpoint}{self.questionnaire.id}', data={})
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-    def test_questionnaire_delete_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.delete(f'{self.base_endpoint}{self.questionnaire.id}')
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-
-@override_settings(ROOT_URLCONF=test_urlconf)
 class TestPublicQuestionEndpoint(ValidateJsonSchemaMixin, APITestCase):
     base_endpoint = '/public/qa/questions/'
 
@@ -172,13 +42,13 @@ class TestPublicQuestionEndpoint(ValidateJsonSchemaMixin, APITestCase):
         self.question = self.questionnaire.first_question
 
         self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_get_question_detail.json')
+            os.path.join(THIS_DIR, '../json_schema/public_get_question_detail.json')
         )
         self.list_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_get_question_list.json')
+            os.path.join(THIS_DIR, '../json_schema/public_get_question_list.json')
         )
         self.post_answer_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_post_question_answer_response.json')
+            os.path.join(THIS_DIR, '../json_schema/public_post_question_answer_response.json')
         )
 
     def test_question_list(self):
@@ -499,140 +369,4 @@ class TestPublicQuestionEndpoint(ValidateJsonSchemaMixin, APITestCase):
 
     def test_answer_question_delete_not_allowed(self):
         response = self.client.delete(f'{self.base_endpoint}{self.question.uuid}/answer')
-        self.assertEqual(response.status_code, 405)
-
-
-@override_settings(ROOT_URLCONF=test_urlconf)
-class TestPrivateQuestionEndpoint(ValidateJsonSchemaMixin, APITestCase):
-    base_endpoint = '/private/qa/questions/'
-
-    def setUp(self):
-        user_model = get_user_model()
-        self.superuser, _ = user_model.objects.get_or_create(
-            email='signals.admin@example.com', is_superuser=True,
-            defaults={'first_name': 'John', 'last_name': 'Doe', 'is_staff': True}
-        )
-        self.questionnaire = QuestionnaireFactory.create()
-        self.question = self.questionnaire.first_question
-
-        self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/private_get_question_detail.json')
-        )
-        self.list_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/private_get_question_list.json')
-        )
-
-    def test_question_list(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.list_schema, response.json())
-
-        self.client.logout()
-
-    def test_question_detail(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.get(f'{self.base_endpoint}{self.question.id}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.detail_schema, response.json())
-
-        self.client.logout()
-
-    def test_question_list_unauthorized(self):
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_question_detail_unauthorized(self):
-        response = self.client.get(f'{self.base_endpoint}{self.question.id}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_question_create_unauthorized(self):
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 401)
-
-    def test_question_update_unauthorized(self):
-        response = self.client.patch(f'{self.base_endpoint}{self.question.id}', data={})
-        self.assertEqual(response.status_code, 401)
-
-    def test_question_delete_unauthorized(self):
-        response = self.client.delete(f'{self.base_endpoint}{self.question.id}')
-        self.assertEqual(response.status_code, 401)
-
-    def test_question_create_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-    def test_question_update_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.patch(f'{self.base_endpoint}{self.question.id}', data={})
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-    def test_question_delete_not_allowed(self):
-        self.client.force_authenticate(user=self.superuser)
-
-        response = self.client.delete(f'{self.base_endpoint}{self.question.id}')
-        self.assertEqual(response.status_code, 405)
-
-        self.client.logout()
-
-
-@override_settings(ROOT_URLCONF=test_urlconf)
-class TestPublicSessionEndpoint(ValidateJsonSchemaMixin, APITestCase):
-    base_endpoint = '/public/qa/sessions/'
-
-    def setUp(self):
-        self.questionnaire = QuestionnaireFactory.create()
-        self.session = SessionFactory.create(questionnaire=self.questionnaire)
-
-        self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, 'json_schema/public_get_session_detail.json')
-        )
-
-    def test_session_detail(self):
-        response = self.client.get(f'{self.base_endpoint}{self.session.uuid}')
-        self.assertEqual(response.status_code, 200)
-
-        self.assertJsonSchema(self.detail_schema, response.json())
-
-    def test_session_detail_gone(self):
-        now = timezone.now()
-        with freeze_time(now - timezone.timedelta(days=1)):
-            session = SessionFactory.create(questionnaire=self.questionnaire)
-
-        response = self.client.get(f'{self.base_endpoint}{session.uuid}')
-        self.assertEqual(response.status_code, 410)
-
-        now = timezone.now()
-        with freeze_time(now - timezone.timedelta(days=1)):
-            session = SessionFactory.create(questionnaire=self.questionnaire,
-                                            submit_before=now-timezone.timedelta(hours=1))
-
-        response = self.client.get(f'{self.base_endpoint}{session.uuid}')
-        self.assertEqual(response.status_code, 410)
-
-    def test_session_list_not_found(self):
-        response = self.client.get(f'{self.base_endpoint}')
-        self.assertEqual(response.status_code, 404)
-
-    def test_session_create_not_found(self):
-        response = self.client.post(f'{self.base_endpoint}', data={})
-        self.assertEqual(response.status_code, 404)
-
-    def test_session_update_not_allowed(self):
-        response = self.client.patch(f'{self.base_endpoint}{self.questionnaire.uuid}', data={})
-        self.assertEqual(response.status_code, 405)
-
-    def test_session_delete_not_allowed(self):
-        response = self.client.delete(f'{self.base_endpoint}{self.questionnaire.uuid}')
         self.assertEqual(response.status_code, 405)
